enhanced-mcp-status-check/requirements.txt
# Enhanced MCP Status Check Requirements

# Core dependencies
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0
pyyaml>=6.0.0
watchdog>=3.0.0

# Authentication dependencies
PyJWT>=2.8.0
cryptography>=41.0.0
aioresponses>=0.7.4

# Performance optimization dependencies
psutil>=5.9.0
numpy>=1.24.0
matplotlib>=3.6.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
aioresponses>=0.7.4

# Development dependencies
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0